requirements.txt
asgiref==3.7.2
Django==4.2.4
django-split-settings==1.2.0
mysqlclient==2.2.0
python-decouple==3.8
sqlparse==0.4.4
